Alterando variaveis para 64 bits

+++ model/ofswitch13-port.cc
@@ -350,7 +350,7 @@ OFSwitch13Port::Receive (Ptr<NetDevice> device, Ptr<const Packet> packet,
   // Check the rate limiter bucket
   if (m_rateLimiter != 0)
     {
-      uint32_t pktSize = packet->GetSize () * 8;
+      uint64_t pktSize = packet->GetSize () * 8;
       if (!m_rateLimiter->removeTokens (pktSize))
         {
           NS_LOG_DEBUG ("Not enough tokens in the bucket to transfer the packet. Discarding");
